requirements.txt
alembic==1.15.1
blinker==1.9.0
click==8.1.8
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-session==0.5.0
greenlet==3.1.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
python-dotenv==1.0.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==2.3.7
k
